Spit Logo components to client and server components
