+++ src/app/sponsors/components/Logos.tsx
@@ -1,61 +1,33 @@
-import generateAtlases, { Atlas, OutputOptions } from "@/lib/atlasLoader";
-import type { SponsorsDatum } from "../types";
+"use client";
+
 import "./styles.css";
+import { Atlas } from "@/lib/atlasLoader";
 import type { Property } from "csstype";
-
-// let sharpObj = new Sharp("@/")
-
-interface LogoAtlasProps {
-	className?: string;
-	logoData: SponsorsDatum[];
-	outputOptions: OutputOptions;
-}
-export default async function Logos({
-	className = "",
-	logoData,
-	outputOptions,
-}: LogoAtlasProps) {
-	const atlases = await generateAtlases(logoData, 4, 400, outputOptions);
-
-	return (
-		<div
-			className={`sponsorLogo ${className}`}
-			style={{ contain: "content", width: "100%" }}
-		>
-			{atlases.map((atlas, index) => (
-				<LogoRow
-					key={index}
-					atlasData={atlas}
-					animationDirection={index % 2 ? "reverse" : "normal"}
-				></LogoRow>
-			))}
-		</div>
-	);
-}
+import { useState } from "react";
 
 interface LogoRowProps {
-	atlasData: Atlas;
-	animationDirection: Property.AnimationDirection;
+  atlasData: Atlas;
+  animationDirection: Property.AnimationDirection;
 }
 
-function LogoRow({
-	atlasData: { atlasFilePath, height, width },
-	animationDirection = "unset",
+export function LogoRow({
+  atlasData: { atlasFilePath, height, width, sprites },
+  animationDirection = "unset",
 }: LogoRowProps) {
-	return (
-		<div
-			className="logoRow"
-			style={{
-				// overflow: "scroll hidden",
-				backgroundImage: `url(${atlasFilePath})`,
-				backgroundSize: "auto 100px",
-				animationDuration: `${(width / height) * 15}s`,
-				animationDirection: animationDirection,
-				backgroundPositionX: `${width}px`,
-			}}
-		>
-		</div>
-	);
+  return (
+    <div
+      className="logoRow"
+      style={{
+        // overflow: "scroll hidden",
+        backgroundImage: `url(${atlasFilePath})`,
+        backgroundSize: "auto 100px",
+        animationDuration: `${(width / height) * 15}s`,
+        animationDirection: animationDirection,
+        backgroundPositionX: `${width}px`,
+      }}
+    >
+    </div>
+  );
 }
 
 interface LogoComponentProps {}
